ussdiface: fix the incorrect use of indexOfMethod()
The argument of indexOfMethod() is supposed to be normallized [1]. The
"const &" shouldn't be around the QString type. Having one make Qt fails
to find the method, always triggering the error code.

It has always been incorrect, but it doesn't cause a problem because the
error handling was incorrect. Commit fc1e09e187a1 ("fix error handling
logic") fixes it, thus exposing this problem.

Fixes https://github.com/HelloVolla/ubuntu-touch-beta-tests/issues/38
Should also fixes the problem on PinePhone as it uses the same code, but
I can't find the filed issue.

[1] https://doc.qt.io/qt-5/qmetaobject.html#indexOfMethod

Acked-by: Alexey Andreyev <aa13q@ya.ru>